Disable a couple of checks added in clang-tidy 8
Will install clang-tidy 8 in CI in following commit.

modernize-avoid-c-arrays is disabled because it is currently not practical
to replace every use of a C array with std::array/vector/string_view. E.g.
there is currently no convenient way to have C friendly string constants
in C++ without using a C array.

modernize-use-nodiscard is too aggressive. It adds a warning for e.g. every
const method that returns a value. This means every simple inline getter
method must be marked with [[nodiscard]]. Better to use [[nodiscard]]
judiciously where it really does matter if the return value is unused and
it is a critical error not to check it. Should reconsider enabling it if
check is modified to not warn for inline getter methods.

readability-magic-numbers suggests adding constants in many cases where it
does not improve readability, e.g.:

EXPECT_EQ(123, string_to_int("123"));

It is possible to list numbers to ignore in configuration but would be nice
if it was possible to say "ignore arguments passed to this function/macro"
so cases like the above could be ignored.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>